scripts/requirements.txt
httpx>=0.27
pyyaml>=6.0
tqdm>=4.66
python-slugify>=8.0
